contact

+++ routes/web.php
@@ -30,9 +30,8 @@ Route::get('/news/{article}', [ArticleController::class, 'show']);
 Route::controller(ContactController::class)->group(function () {
     Route::get('/contact/index', 'index');
     Route::post('/contactform', 'store');
-    Route::get('/contact/{contact}', 'show')->middleware('auth');
-    Route::put('/contact/{contact}', 'update');
-    Route::delete('/contact/{contact}', 'destroy');
+    Route::get('/contact/{contact}', 'show')->middleware(['auth',isAdmin::class]);
+    Route::delete('/contact/{contact}', 'destroy')->middleware(['auth',isAdmin::class]);
 });
 Route::middleware(['auth',isAdmin::class])->group(function () {
     Route::get('/dashboard', function () {
